Create Plugin superclass, and make Output inherit from it. Fix LDR max.
Create a new 'Plugin' object from which specific plugins can inherit.
Change the Output object to inherit from the new Plugin object. Finally
fix the LDR at full voltage issue (hopefully!).
